Require count limit to be positive
This patch changes the field type to be a PositiveSmallInteger. The
serializer is updated to give a nice error when user tries to set
negative value. A test for this feature is added.

+++ pdc/apps/contact/models.py
@@ -16,8 +16,10 @@ from django.utils.translation import ugettext_lazy as _
 class ContactRole(models.Model):
 
     name = models.CharField(max_length=128, unique=True)
-    count_limit = models.IntegerField(default=1,
-                                      help_text=_('Contact count limit of the role for each component.'))
+    count_limit = models.PositiveSmallIntegerField(
+        default=1,
+        help_text=_('Contact count limit of the role for each component.')
+    )
     UNLIMITED = 0
 
     def _get_max_component_role_count(self, component_model, role):
